List view working again
List editor + list table editor not working yet

+++ RapidCMS/TestLibrary/TestEntity.cs
@@ -24,7 +24,7 @@ namespace TestLibrary
 
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Number { get; set; }
+        public long Number { get; set; }
     }
 
     public class TestEntityVariantA : TestEntity
